[#6] 부저 모듈 tcp 통신
