--- CLIENT/buzzer/receive/.buzzer.ko.cmd
@@ -1,1 +1,0 @@
-cmd_/home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.ko := /usr/local/arm/4.4.1/bin/arm-none-linux-gnueabi-ld -EL -r  -T /home/mds/exercise/linux/teacher/linux-3.0.22/scripts/module-common.lds --build-id  -o /home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.ko /home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.o /home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.mod.o
