--- CLIENT/buzzer/receive/.tmp_versions/buzzer.mod
@@ -1,2 +1,0 @@
-/home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.ko
-/home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.o
