--- CLIENT/buzzer/receive/modules.order
@@ -1,1 +1,0 @@
-kernel//home/mds/git/rc_car/SERVER/device/buzzer/receive/buzzer.ko
