+++ CLIENT/rc_client/Makefile
@@ -15,22 +15,24 @@
 #  $  export ARCH=x86 ; make
 #
 
+obj-m := buzzer.o
 
-ifeq ("$(ARCH)", "arm")
-	CROSS_COMPILE=arm-none-linux-gnueabi-
-endif
+CROSS_COMPILE=arm-none-linux-gnueabi-
 
 KDIR := /work/linux
 
 allc:
-	$(CROSS_COMPILE)gcc -o example example.c   `sdl-config-$(ARCH) --cflags --libs`   -lSDL -lSDL_ttf -lSDL_image -lpthread 
+	make -C $(KDIR) SUBDIRS=$(PWD) modules 
+	$(CROSS_COMPILE)gcc -o example example.c th_buzzer.c  `sdl-config-$(ARCH) --cflags --libs`   -lSDL -lSDL_ttf -lSDL_image -lpthread 
 
 allcpp:
 	$(CROSS_COMPILE)g++ -o example example.cpp `sdl-config-$(ARCH) --cflags --libs`   -lSDL -lSDL_ttf -lSDL_image  
 install:
 	cp  example  /nfsroot/root
+	cp 	buzzer.ko /nfsroot/root
 
 clean:
+	make -C $(KDIR) SUBDIRS=$(PWD) clean 
 	rm -rf example
 	rm -rf *.o
- 
+
