+++ CLIENT/rc_client/th_buzzer.h
@@ -1,0 +1,11 @@
+#ifndef __TH_BUZZER_H__
+#define __TH_BUZZER_H__
+
+#define	BUZZER_SIG           6
+#define DEV_BUZZER_NAME  	"buzzer"
+#define	DEV_BUZZER_MAJOR 	102
+
+void buzzer_mknod(void);
+void *buzzer_func(void *data);
+
+#endif //__TH_BUZZER_H__
